// SGUEES-SPA/src/app/pages/Academics/aca-bec-origen-beca/models/aca-bec-origen-beca.ts
export interface AcaBecOrigenBeca {
	CORR_EMPRESA: number;
	CORR_ORIGEN_BECA: number;
	CODIGO_ORIGEN: string;
	NOMBRE_ORIGEN: string;
	DESCRIPCION: string;
	ACTIVO: boolean;
	USUARIO_CREA: string;
	ESTACION_CREA: string;
	FECHA_CREA: Date;
	USUARIO_ACTU: string;
	ESTACION_ACTU: string;
	FECHA_ACTU: Date;
}